+++ desktop/integration-test/jest.config.json
@@ -1,8 +1,8 @@
 {
+  "globalSetup": "<rootDir>/build.ts",
   "transform": {
     "\\.[jt]sx?$": ["babel-jest", { "rootMode": "upward" }]
   },
-  "testPathIgnorePatterns": ["integration-test"],
   "//": "Native find is slow because it does not exclude files: https://github.com/facebook/jest/pull/11264#issuecomment-825377579",
   "haste": { "forceNodeFilesystemAPI": true }
 }
